use parse-duration to parse the estimated time of a task

+++ src/interface.rs
@@ -1,6 +1,5 @@
 use std::path::PathBuf;
 use rusqlite::{params, Connection, Result};
-use crate::tasks;
 
 pub fn init_journal(journal_path: PathBuf) -> Result<()> {
     let conn = Connection::open(journal_path)?;
@@ -21,16 +20,14 @@ pub fn init_journal(journal_path: PathBuf) -> Result<()> {
     Ok(())
 }
 
-pub fn add_task(journal_path: PathBuf, description: String, estimated_time: usize) -> Result<()> {
+pub fn add_task(journal_path: PathBuf, description: String, estimated_time: u64) -> Result<()> {
     let conn = Connection::open(journal_path)?;
 
     //let mut stmt = conn.prepare()?;
-    let position = conn.query_row("SELECT count(*) from task where day = DATE()", [], |row| row.get::<_,i32>(0))?;
+    let position = conn.query_row("SELECT count(*) from task where day = DATE()", [], |row| row.get::<_,usize>(0))?;
     //println!(position);
 
     conn.execute("INSERT INTO task (day, description, position, created_at, estimated_time)
 VALUES(CURRENT_DATE, ?1, ?2, CURRENT_TIME, ?3)", params![description, position, estimated_time])?;
     Ok(())
 }
-//pub fn complete_task(journal_path: PathBuf, task_position: usize) -> Result<()> { ... }
-//pub fn list_tasks(journal_path: PathBuf) -> Result<()> { ... }
